python-backend/requirements.txt
Pillow==8.1.0
PyYAML==5.4.1
websockets==8.1
